% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_clean_tournament_dataframe.R
\name{odds_clean_tournament_dataframe}
\alias{odds_clean_tournament_dataframe}
\title{Clean the dataframe downloaded from \code{oddsportal}}
\usage{
odds_clean_tournament_dataframe(oddsportal_raw)
}
\arguments{
\item{oddsportal_raw}{raw dataframe from \code{odds_table_from_url()}}
}
\description{
Clean the dataframe downloaded from \code{oddsportal}
}
