% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_set_up_virtual_browser.R
\name{odds_set_up_virtual_browser}
\alias{odds_set_up_virtual_browser}
\title{set up the virtual browser}
\usage{
odds_set_up_virtual_browser(port = 8912L)
}
\arguments{
\item{port}{an integer or numeric 4 digit input}
}
\description{
set up the virtual browser
}
