% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_table_from_url.R
\name{odds_table_from_url}
\alias{odds_table_from_url}
\title{A function top download tournament results table from \code{oddsportal}}
\usage{
odds_table_from_url(url)
}
\arguments{
\item{url}{a base text url leading to \code{oddsportal} results table}
}
\description{
The function takes \code{oddsportal} url as input and downloads the tournament data.
}
